Prepare security-manager for master/slave mode
This commit prepares security-manager for work in master/slave mode.
In order to properly install/uninstall applications inside containers,
security-manager inside container (slave) must delegate calls related to
Cynara/SMACK to security-manager outside a container (master).
Since entire master/slave mode is a huge change, it is divided into two
commits - this is the first one. Logic for master service and changes in
service to work as slave are left for second commit.

With this change security-manager launched without additional arguments should
work as it did.

Change-Id: If05cdeb2d2c35c046bf4cb46d884a3689dab57ad
Signed-off-by: Lukasz Kostyra <l.kostyra@samsung.com>