Added examination of error revents in poll
poll may return positive value even when error condition occured and
the event was not requested for. It applies to events like: POLLNVAL,
POLLERR and POLLHUP. Since no checking of the events is performed it is
possible to continue for the server with an error condition that can be
caught somewhere later.
This patch examinates these error events and changes the return value of
waitForSocket function to -1 to indicate an error as quick as possbile.
Of course, this could also be done by upper calls that would decide was
is the best thing to do for each of the error cases.

Change-Id: Ib6ede6ab228443305b1b1ce943d5ca84eedebfcf
Signed-off-by: Sebastian Grabowski <s.grabowski@samsung.com>