Add installation types (global, local, preloaded).
Before this commit installation type was based on UID.
With this commit it is possible to set type of installation (global, local,
preloaded) during app installation request. If type is not specified,
and installation is performed by global user, default 'global' type of
installation is set. Otherwise installation type is set to 'local'.

New API function avaliable:

* int security_manager_app_inst_req_set_install_type(app_inst_req *p_req,
    const enum app_install_type &type)

Change-Id: I745da8fc7a7393c360ed6d281a1f729d22bb89e6