Move author_id to pkg - server code adjustment
[Problem] Author is not a feature of app anymore. Server code needs to be
adjusted.
[Solution] Get author via pkg instead of app. Rename variables and functions.
Update author's rules in existing ones if a new app with different author is
installed. Separate author rules for app from app-rules-template.smack

[Verification] Run tests (especially author related ones)
TODO: The test 46 fails because it assumes that a trusted path is owned by an
app. Now it's owned by the pkg and common smack rules won't be revoked until
all pkg apps are uninstalled.

Change-Id: I8e42877170809e9e71c8c676b566119e3b16fbd5