Fix for a duplicate SharedRO rule part 2
This fix prevents creation of the duplicate SharedRO rules for already
installed Tizen 2.X application when installing a new Tizen 2.X
application.

It works by modifying the SQL query in a way that it's possible to
exclude apps that originate in the same package as the app that's being
installed.

Change-Id: Id87c729eaf9170e968cb701b7327ad62b18c59a8